Update DCF77.h
- added function with return of CEST: getCEST()
returns 1 if CEST and 0 if CET (MEZ).

- added delays for filtering short PRE/POST-pulses (see DCF77.h)

- changed functions for rPW and rCT to "-" for Prepulses and "+" for Postpulses
i have changed the functions which reject short pulses before or after a correct pulse. It seems to correct some of that short jitter pulses better.They are detected in debug mode, but nevertheless in some cases the buffer is good to be used.

-addes some function names for getWeekday() (to be implemented), 
getEncWeatherData() to get the encrypted MeteoTime-Data for (external) decryption,
and getAlertData() to get the Civil Alert Data.